started RepairerJPARepository

+++ src/main/java/org/example/repository/JdbcRepositiries/OrderJDBCRepository.java
@@ -85,7 +85,7 @@ public class OrderJDBCRepository implements OrderRepository {
             while (resultSet.next()) {
                 Repairer rep = Repairer.builder()
                         .name(resultSet.getString("name"))
-                        .id(resultSet.getInt("id"))
+                        .id(resultSet.getLong("id"))
                         .status(RepairerStatus.valueOf(resultSet.getString("status")))
                         .build();
                 repairers.add(rep);
